requirements.txt
streamlit 
google-generativeai
python-dotenv
python -m streamlit run app.py